refactor: deleted the "show-notes" component, and created the "note.js" component, and changed the logic into the UserProfile view.
